requirements.txt
streamlit
pandas
xlsxwriter
holidays
python-docx